+++ src/server/worker/api_server.go
@@ -10,10 +10,8 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
-	"os/user"
 	"path"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -46,7 +44,6 @@ import (
 	"github.com/pachyderm/pachyderm/src/server/pkg/backoff"
 	col "github.com/pachyderm/pachyderm/src/server/pkg/collection"
 	"github.com/pachyderm/pachyderm/src/server/pkg/errutil"
-	"github.com/pachyderm/pachyderm/src/server/pkg/exec"
 	"github.com/pachyderm/pachyderm/src/server/pkg/hashtree"
 	"github.com/pachyderm/pachyderm/src/server/pkg/obj"
 	"github.com/pachyderm/pachyderm/src/server/pkg/ppsdb"
@@ -56,15 +53,11 @@ import (
 	"github.com/pachyderm/pachyderm/src/server/pkg/watch"
 	"github.com/pachyderm/pachyderm/src/server/worker/common"
 	"github.com/pachyderm/pachyderm/src/server/worker/logs"
+	"github.com/pachyderm/pachyderm/src/server/worker/stats"
 )
 
 const (
-	// The maximum number of concurrent download/upload operations
-	concurrency = 100
-
 	planPrefix        = "/plan"
-	chunkPrefix       = "/chunk"
-	mergePrefix       = "/merge"
 	shardPrefix       = "/shard"
 	shardTTL          = 30
 	noShard           = int64(-1)
@@ -117,8 +110,6 @@ type APIServer struct {
 	namespace string
 	// The jobs collection
 	jobs col.Collection
-	// The pipelines collection
-	pipelines col.Collection
 	// The plans collection
 	// Stores chunk layout and merges
 	plans col.Collection
@@ -132,9 +123,6 @@ type APIServer struct {
 
 	// We only export application statistics if enterprise is enabled
 	exportStats bool
-
-	uid *uint32
-	gid *uint32
 
 	// hashtreeStorage is the where we store on disk hashtrees
 	hashtreeStorage string
@@ -184,7 +172,6 @@ func NewAPIServer(pachClient *client.APIClient, etcdClient *etcd.Client, etcdPre
 		workerName:      workerName,
 		namespace:       namespace,
 		jobs:            ppsdb.Jobs(etcdClient, etcdPrefix),
-		pipelines:       ppsdb.Pipelines(etcdClient, etcdPrefix),
 		plans:           col.NewCollection(etcdClient, path.Join(etcdPrefix, planPrefix), nil, &Plan{}, nil, nil),
 		shards:          col.NewCollection(etcdClient, path.Join(etcdPrefix, shardPrefix, pipelineInfo.Pipeline.Name), nil, &ShardInfo{}, nil, nil),
 		hashtreeStorage: hashtreeStorage,
@@ -244,33 +231,11 @@ func NewAPIServer(pachClient *client.APIClient, etcdClient *etcd.Client, etcdPre
 		}
 		if server.pipelineInfo.Transform.Cmd == nil {
 			if len(image.Config.Entrypoint) == 0 {
-				ppsutil.FailPipeline(ctx, etcdClient, server.pipelines,
+				ppsutil.FailPipeline(ctx, etcdClient, ppsdb.Pipelines(etcdClient, etcdPrefix),
 					pipelineInfo.Pipeline.Name,
 					"nothing to run: no transform.cmd and no entrypoint")
 			}
 			server.pipelineInfo.Transform.Cmd = image.Config.Entrypoint
-		}
-	}
-	if pipelineInfo.Transform.User != "" {
-		user, err := lookupDockerUser(pipelineInfo.Transform.User)
-		if err != nil && !os.IsNotExist(err) {
-			return nil, err
-		}
-		// If `user` is `nil`, `uid` and `gid` will get set, and we won't
-		// customize the user that executes the worker process.
-		if user != nil { // user is nil when os.IsNotExist(err) is true in which case we use root
-			uid, err := strconv.ParseUint(user.Uid, 10, 32)
-			if err != nil {
-				return nil, err
-			}
-			uid32 := uint32(uid)
-			server.uid = &uid32
-			gid, err := strconv.ParseUint(user.Gid, 10, 32)
-			if err != nil {
-				return nil, err
-			}
-			gid32 := uint32(gid)
-			server.gid = &gid32
 		}
 	}
 	switch {
@@ -285,213 +250,27 @@ func NewAPIServer(pachClient *client.APIClient, etcdClient *etcd.Client, etcdPre
 	return server, nil
 }
 
-func (a *APIServer) downloadGitData(pachClient *client.APIClient, dir string, input *common.Input) error {
-	file := input.FileInfo.File
-	pachydermRepoName := input.Name
-	var rawJSON bytes.Buffer
-	err := pachClient.GetFile(pachydermRepoName, file.Commit.ID, "commit.json", 0, 0, &rawJSON)
-	if err != nil {
-		return err
-	}
-	var payload github.PushPayload
-	err = json.Unmarshal(rawJSON.Bytes(), &payload)
-	if err != nil {
-		return err
-	}
-	sha := payload.After
-	// Clone checks out a reference, not a SHA
-	r, err := git.PlainClone(
-		filepath.Join(dir, pachydermRepoName),
-		false,
-		&git.CloneOptions{
-			URL:           payload.Repository.CloneURL,
-			SingleBranch:  true,
-			ReferenceName: gitPlumbing.ReferenceName(payload.Ref),
-		},
-	)
-	if err != nil {
-		return fmt.Errorf("error cloning repo %v at SHA %v from URL %v: %v", pachydermRepoName, sha, input.GitURL, err)
-	}
-	hash := gitPlumbing.NewHash(sha)
-	wt, err := r.Worktree()
-	if err != nil {
-		return err
-	}
-	err = wt.Checkout(&git.CheckoutOptions{Hash: hash})
-	if err != nil {
-		return fmt.Errorf("error checking out SHA %v from repo %v: %v", sha, pachydermRepoName, err)
-	}
-	return nil
-}
-
-func (a *APIServer) downloadData(pachClient *client.APIClient, logger logs.TaggedLogger, inputs []*common.Input, puller *filesync.Puller, stats *pps.ProcessStats, statsTree *hashtree.Ordered) (_ string, retErr error) {
-	defer a.reportDownloadTimeStats(time.Now(), stats, logger)
-	logger.Logf("starting to download data")
-	defer func(start time.Time) {
-		if retErr != nil {
-			logger.Logf("errored downloading data after %v: %v", time.Since(start), retErr)
-		} else {
-			logger.Logf("finished downloading data after %v", time.Since(start))
-		}
-	}(time.Now())
-	dir := filepath.Join(client.PPSInputPrefix, client.PPSScratchSpace, uuid.NewWithoutDashes())
-	// Create output directory (currently /pfs/out)
-	outPath := filepath.Join(dir, "out")
-	if a.pipelineInfo.Spout != nil {
-		// Spouts need to create a named pipe at /pfs/out
-		if err := os.MkdirAll(filepath.Dir(outPath), 0700); err != nil {
-			return "", fmt.Errorf("mkdirall :%v", err)
-		}
-		if err := syscall.Mkfifo(outPath, 0666); err != nil {
-			return "", fmt.Errorf("mkfifo :%v", err)
-		}
-	} else {
-		if err := os.MkdirAll(outPath, 0777); err != nil {
-			return "", err
-		}
-	}
-	for _, input := range inputs {
-		if input.GitURL != "" {
-			if err := a.downloadGitData(pachClient, dir, input); err != nil {
-				return "", err
-			}
-			continue
-		}
-		file := input.FileInfo.File
-		root := filepath.Join(dir, input.Name, file.Path)
-		var statsRoot string
-		if statsTree != nil {
-			statsRoot = path.Join(input.Name, file.Path)
-			parent, _ := path.Split(statsRoot)
-			statsTree.MkdirAll(parent)
-		}
-		if err := puller.Pull(pachClient, root, file.Commit.Repo.Name, file.Commit.ID, file.Path, input.Lazy, input.EmptyFiles, concurrency, statsTree, statsRoot); err != nil {
-			return "", err
-		}
-	}
-	return dir, nil
-}
-
-func (a *APIServer) linkData(inputs []*common.Input, dir string) error {
-	// Make sure that previously symlinked outputs are removed.
-	if err := a.unlinkData(inputs); err != nil {
-		return err
-	}
-	for _, input := range inputs {
-		src := filepath.Join(dir, input.Name)
-		dst := filepath.Join(client.PPSInputPrefix, input.Name)
-		if err := os.Symlink(src, dst); err != nil {
-			return err
-		}
-	}
-	return os.Symlink(filepath.Join(dir, "out"), filepath.Join(client.PPSInputPrefix, "out"))
-}
-
-func (a *APIServer) unlinkData(inputs []*common.Input) error {
-	dirs, err := ioutil.ReadDir(client.PPSInputPrefix)
-	if err != nil {
-		return fmt.Errorf("ioutil.ReadDir: %v", err)
-	}
-	for _, d := range dirs {
-		if d.Name() == client.PPSScratchSpace {
-			continue // don't delete scratch space
-		}
-		if err := os.RemoveAll(filepath.Join(client.PPSInputPrefix, d.Name())); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// Run user error code and return the combined output of stdout and stderr.
-func (a *APIServer) runUserErrorHandlingCode(ctx context.Context, logger logs.TaggedLogger, environ []string, stats *pps.ProcessStats, rawDatumTimeout *types.Duration) (retErr error) {
-	logger.Logf("beginning to run user error handling code")
-	defer func(start time.Time) {
-		if retErr != nil {
-			logger.Logf("errored running user error handling code after %v: %v", time.Since(start), retErr)
-		} else {
-			logger.Logf("finished running user error handling code after %v", time.Since(start))
-		}
-	}(time.Now())
-
-	cmd := exec.CommandContext(ctx, a.pipelineInfo.Transform.ErrCmd[0], a.pipelineInfo.Transform.ErrCmd[1:]...)
-	if a.pipelineInfo.Transform.ErrStdin != nil {
-		cmd.Stdin = strings.NewReader(strings.Join(a.pipelineInfo.Transform.ErrStdin, "\n") + "\n")
-	}
-	cmd.Stdout = logger.WithUserCode()
-	cmd.Stderr = logger.WithUserCode()
-	cmd.Env = environ
-	if a.uid != nil && a.gid != nil {
-		cmd.SysProcAttr = &syscall.SysProcAttr{
-			Credential: &syscall.Credential{
-				Uid: *a.uid,
-				Gid: *a.gid,
-			},
-		}
-	}
-	cmd.Dir = a.pipelineInfo.Transform.WorkingDir
-	err := cmd.Start()
-	if err != nil {
-		return fmt.Errorf("error cmd.Start: %v", err)
-	}
-	// A context w a deadline will successfully cancel/kill
-	// the running process (minus zombies)
-	state, err := cmd.Process.Wait()
-	if err != nil {
-		return fmt.Errorf("error cmd.Wait: %v", err)
-	}
-	if common.isDone(ctx) {
-		if err = ctx.Err(); err != nil {
-			return err
-		}
-	}
-	// Because of this issue: https://github.com/golang/go/issues/18874
-	// We forked os/exec so that we can call just the part of cmd.Wait() that
-	// happens after blocking on the process. Unfortunately calling
-	// cmd.Process.Wait() then cmd.Wait() will produce an error. So instead we
-	// close the IO using this helper
-	err = cmd.WaitIO(state, err)
-	// We ignore broken pipe errors, these occur very occasionally if a user
-	// specifies Stdin but their process doesn't actually read everything from
-	// Stdin. This is a fairly common thing to do, bash by default ignores
-	// broken pipe errors.
-	if err != nil && !strings.Contains(err.Error(), "broken pipe") {
-		// (if err is an acceptable return code, don't return err)
-		if exiterr, ok := err.(*exec.ExitError); ok {
-			if status, ok := exiterr.Sys().(syscall.WaitStatus); ok {
-				for _, returnCode := range a.pipelineInfo.Transform.AcceptReturnCode {
-					if int(returnCode) == status.ExitStatus() {
-						return nil
-					}
-				}
-			}
-		}
-		return fmt.Errorf("error cmd.WaitIO: %v", err)
-	}
-	return nil
-}
-
 func (a *APIServer) reportUploadStats(start time.Time, stats *pps.ProcessStats, logger logs.TaggedLogger) {
 	duration := time.Since(start)
 	stats.UploadTime = types.DurationProto(duration)
 	if a.exportStats {
-		if hist, err := datumUploadTime.GetMetricWithLabelValues(a.pipelineInfo.ID, a.jobID); err != nil {
-			logger.Logf("failed to get histogram w labels: pipeline (%v) job (%v) with error %v", a.pipelineInfo.ID, a.jobID, err)
+		if hist, err := stats.DatumUploadTime.GetMetricWithLabelValues(a.pipelineInfo.ID, logger.JobID()); err != nil {
+			logger.Logf("failed to get histogram w labels: pipeline (%v) job (%v) with error %v", a.pipelineInfo.ID, logger.JobID(), err)
 		} else {
 			hist.Observe(duration.Seconds())
 		}
-		if counter, err := datumUploadSecondsCount.GetMetricWithLabelValues(a.pipelineInfo.ID, a.jobID); err != nil {
-			logger.Logf("failed to get counter w labels: pipeline (%v) job (%v) with error %v", a.pipelineInfo.ID, a.jobID, err)
+		if counter, err := stats.DatumUploadSecondsCount.GetMetricWithLabelValues(a.pipelineInfo.ID, logger.JobID()); err != nil {
+			logger.Logf("failed to get counter w labels: pipeline (%v) job (%v) with error %v", a.pipelineInfo.ID, logger.JobID(), err)
 		} else {
 			counter.Add(duration.Seconds())
 		}
-		if hist, err := datumUploadSize.GetMetricWithLabelValues(a.pipelineInfo.ID, a.jobID); err != nil {
-			logger.Logf("failed to get histogram w labels: pipeline (%v) job (%v) with error %v", a.pipelineInfo.ID, a.jobID, err)
+		if hist, err := stats.DatumUploadSize.GetMetricWithLabelValues(a.pipelineInfo.ID, logger.JobID()); err != nil {
+			logger.Logf("failed to get histogram w labels: pipeline (%v) job (%v) with error %v", a.pipelineInfo.ID, logger.JobID(), err)
 		} else {
 			hist.Observe(float64(stats.UploadBytes))
 		}
-		if counter, err := datumUploadBytesCount.GetMetricWithLabelValues(a.pipelineInfo.ID, a.jobID); err != nil {
-			logger.Logf("failed to get counter w labels: pipeline (%v) job (%v) with error %v", a.pipelineInfo.ID, a.jobID, err)
+		if counter, err := stats.DatumUploadBytesCount.GetMetricWithLabelValues(a.pipelineInfo.ID, logger.JobID()); err != nil {
+			logger.Logf("failed to get counter w labels: pipeline (%v) job (%v) with error %v", a.pipelineInfo.ID, logger.JobID(), err)
 		} else {
 			counter.Add(float64(stats.UploadBytes))
 		}
@@ -769,7 +548,8 @@ func (a *APIServer) datum() []*pps.InputFile {
 	return result
 }
 
-func (a *APIServer) userCodeEnv(jobID string, outputCommitID string, data []*common.Input) []string {
+// TODO: move this to common
+func userCodeEnv(jobID string, outputCommitID string, data []*common.Input) []string {
 	result := os.Environ()
 	for _, input := range data {
 		result = append(result, fmt.Sprintf("%s=%s", input.Name, filepath.Join(client.PPSInputPrefix, input.Name, input.FileInfo.File.Path)))
@@ -1289,7 +1069,7 @@ func (a *APIServer) cancelCtxIfJobFails(jobCtx context.Context, jobCancel func()
 
 	backoff.RetryNotify(func() error {
 		// Check if job was cancelled while backoff was sleeping
-		if common.isDone(jobCtx) {
+		if common.IsDone(jobCtx) {
 			return nil
 		}
 
@@ -1621,11 +1401,10 @@ func (a *APIServer) processDatums(pachClient *client.APIClient, logger logs.Tagg
 				}()
 			}
 
-			env := a.userCodeEnv(jobInfo.Job.ID, jobInfo.OutputCommit.ID, data)
 			var dir string
 			var failures int64
 			if err := backoff.RetryNotify(func() error {
-				if common.isDone(ctx) {
+				if common.IsDone(ctx) {
 					return ctx.Err() // timeout or cancelled job--don't run datum
 				}
 
@@ -1642,12 +1421,13 @@ func (a *APIServer) processDatums(pachClient *client.APIClient, logger logs.Tagg
 					a.stats = stats
 				}()
 
-				subStats, err := utils.WithProvisionedNode(, func(subStats *pps.ProcessStats) error {
+				subStats, err := utils.WithProvisionedNode(ctx, logger, func(subStats *pps.ProcessStats) error {
+					env := userCodeEnv(jobInfo.Job.ID, jobInfo.OutputCommit.ID, data)
 					a.runMu.Lock()
 					defer a.runMu.Unlock()
-					if err := a.runUserCode(ctx, logger, env, subStats, jobInfo.DatumTimeout); err != nil {
+					if err := utils.runUserCode(ctx, logger, env, subStats, jobInfo.DatumTimeout); err != nil {
 						if a.pipelineInfo.Transform.ErrCmd != nil && failures == jobInfo.DatumTries-1 {
-							if err = a.runUserErrorHandlingCode(ctx, logger, env, subStats, jobInfo.DatumTimeout); err != nil {
+							if err = utils.runUserErrorHandlingCode(ctx, logger, env, subStats, jobInfo.DatumTimeout); err != nil {
 								return fmt.Errorf("error runUserErrorHandlingCode: %v", err)
 							}
 							return errDatumRecovered
@@ -1661,7 +1441,7 @@ func (a *APIServer) processDatums(pachClient *client.APIClient, logger logs.Tagg
 
 				return a.uploadOutput(pachClient, dir, tag, logger, data, subStats, outputTree, datumIdx)
 			}, &backoff.ZeroBackOff{}, func(err error, d time.Duration) error {
-				if common.isDone(ctx) {
+				if common.IsDone(ctx) {
 					return ctx.Err() // timeout or cancelled job, err out and don't retry
 				}
 				failures++
@@ -1886,80 +1666,4 @@ func (a *APIServer) mergeChunk(logger logs.TaggedLogger, high int64, result *pro
 		return a.chunkStatsCache.Put(high, buf)
 	}
 	return nil
-}
-
-// lookupDockerUser looks up users given the argument to a Dockerfile USER directive.
-// According to Docker's docs this directive looks like:
-// USER <user>[:<group>] or
-// USER <UID>[:<GID>]
-func lookupDockerUser(userArg string) (_ *user.User, retErr error) {
-	userParts := strings.Split(userArg, ":")
-	userOrUID := userParts[0]
-	groupOrGID := ""
-	if len(userParts) > 1 {
-		groupOrGID = userParts[1]
-	}
-	passwd, err := os.Open("/etc/passwd")
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		if err := passwd.Close(); err != nil && retErr == nil {
-			retErr = err
-		}
-	}()
-	scanner := bufio.NewScanner(passwd)
-	for scanner.Scan() {
-		parts := strings.Split(scanner.Text(), ":")
-		if parts[0] == userOrUID || parts[2] == userOrUID {
-			result := &user.User{
-				Username: parts[0],
-				Uid:      parts[2],
-				Gid:      parts[3],
-				Name:     parts[4],
-				HomeDir:  parts[5],
-			}
-			if groupOrGID != "" {
-				if parts[0] == userOrUID {
-					// groupOrGid is a group
-					group, err := lookupGroup(groupOrGID)
-					if err != nil {
-						return nil, err
-					}
-					result.Gid = group.Gid
-				} else {
-					// groupOrGid is a gid
-					result.Gid = groupOrGID
-				}
-			}
-			return result, nil
-		}
-	}
-	if err := scanner.Err(); err != nil {
-		log.Fatal(err)
-	}
-	return nil, fmt.Errorf("user %s not found", userArg)
-}
-
-func lookupGroup(group string) (_ *user.Group, retErr error) {
-	groupFile, err := os.Open("/etc/group")
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		if err := groupFile.Close(); err != nil && retErr == nil {
-			retErr = err
-		}
-	}()
-	scanner := bufio.NewScanner(groupFile)
-	for scanner.Scan() {
-		parts := strings.Split(scanner.Text(), ":")
-		if parts[0] == group {
-			return &user.Group{
-				Gid:  parts[2],
-				Name: parts[0],
-			}, nil
-		}
-	}
-	return nil, fmt.Errorf("group %s not found", group)
 }
