+++ src/server/worker/stats/stats_test.go
@@ -1,4 +1,4 @@
-package worker
+package stats
 
 import (
 	"fmt"
